\name{Exported-typos}
\alias{Exported-typos}
\alias{nameS}
\alias{nameS<-}
\alias{typoef}
\alias{geom_warp}
\alias{fitler}
\alias{transmutate}
\alias{lenght}
\title{List of currently exported typos}
\description{
List of currently exported typos
}
\details{
Exported typos:
\itemize{
\item \code{nameS <- .typo(names)}
\item \code{`nameS<-` <- .typo(`names<-`)}
\item \code{typoef <- .typo(typeof)}
\item \code{geom_warp <- .typo(geom_wrap,ggplot2)}
\item \code{fitler <- .typo(filter,dplyr)}
\item \code{transmutate <- .typo(mutate,dplyr)}
\item \code{lenght <- .typo(length)}
}
}

