% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gen_typo.R
\name{Gen_typo}
\alias{Gen_typo}
\title{Generate Typo Documentation}
\usage{
Gen_typo(
  incorrect_spellings,
  correct_spelling,
  methods = NULL,
  pkg = "base",
  dir = "R/"
)
}
\arguments{
\item{incorrect_spellings}{The new incorrect spelling(s) of the function}

\item{correct_spelling}{The correct spelling of the function}

\item{methods}{Additional methods to be implemented.
Possible values include \code{"<-"}}

\item{dir}{Directory for the generated file to be stored}
}
\description{
Generates Generic Typo Documentation
}
\examples{
Gen_typo("nameS","names",methods = c("<-"))
Gen_typo("typoef","typeof")
}
