% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/names.R
\name{names-typo}
\alias{names-typo}
\alias{nameS}
\alias{nameS<-}
\title{Typo of the \code{names()} function}
\usage{
nameS(...)

nameS(...) <- value
}
\arguments{
\item{...}{arguments passed to \code{\link[base:names]{base::names()}}}
}
\description{
Runs the \code{\link[base:names]{base::names()}} function when \code{nameS()} is called and outputs
a warning to the user to alert them to the typo
}
\seealso{
\code{\link[base:names]{base::names()}}
}
