% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/typeof.R
\name{typeof-typo}
\alias{typeof-typo}
\alias{typoef}
\title{Typo of the \code{typeof()} function}
\usage{
typoef(...)
}
\arguments{
\item{...}{arguments passed to \code{\link[base:typeof]{base::typeof()}}}
}
\description{
Runs the \code{\link[base:typeof]{base::typeof()}} function when \code{typoef()} is called and outputs
a warning to the user to alert them to the typo
}
\seealso{
\code{\link[base:typeof]{base::typeof()}}
}
