% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/typo-functions.R
\name{typo-functions}
\alias{typo-functions}
\alias{.typo}
\alias{.typo_alert}
\title{Defining typos}
\usage{
.typo(.correct, .package = base, .typo_function = .typo_alert)

.typo_alert(.correct, .call)
}
\arguments{
\item{.correct}{The correctly spelled function for which the wrapper is being defined}

\item{.package}{The name of the package containing the correct function}

\item{.typo_function}{Function to be used as an alert}

\item{.call}{The incorrectly spelled call made by the user}
}
\value{
Wrapper function to call the correctly spelled function
}
\description{
Function used to create a typo and alert the user
}
\details{
The \code{.typo()} function is used to define a typo within the context
of the \code{typo} package. Without \code{typo}, mistakenly typed functions
will cause an Error and not be evaluated; however with \code{typo},
the function \emph{is} evaluated and a Warning is thrown.
}
\examples{
nameS <- .typo(names)

.typo_alert("names","nameS(mtcars)")

}
