% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/typos.R
\docType{package}
\name{typos}
\alias{typos}
\title{typos: Generate and use typo functions}
\description{
A package designed to help users beat common typos
}
\details{
This package is mostly powered by the \code{\link[=.typo]{.typo()}} function
which provides a flexible and easy to use method
of declaring common typing errors.
}
